+++ packages/stylelint/src/uniapp.ts
@@ -3,10 +3,4 @@ import { MINIAPP_SHARED_CONFIG } from './shared/miniapp'
 
 export default {
   ...MINIAPP_SHARED_CONFIG,
-  overrides: [
-    {
-      files: ['*.wxss'],
-      customSyntax: 'postcss',
-    },
-  ],
 } satisfies Config